requirements.txt
requests==2.31
aiogram==3.3
pyyaml==6.0.1
python-dotenv==1.0.1
